Sign. I was commiting to the wrong repo :(
I was in a different repo on my terminal
and so nothing was actually happening.

Here are a list of changes:

Improved prompt_user function to handle
all edge cases I know of so far, including
empty input and positive/negative values.

Used this improved function for exercise 4

Completed exercise 11 and 4