Fix missing virtual DB
When a db is loaded,
a list of Employees is created which
represents the employee db contents.

A virtual db is created to represent changes to
the db contents, and is whats accessed and
eventually written to the real db.

A virtual db is not created when no exisiting db
file exists yet, which causes an error
on access or write attempts.

So I added one